pyproject.toml
[project]
name = "map-plot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "osmpythontools>=0.3.5",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "svg-py>=1.6.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.0",
]
